Changed hook handlers to synthetic states
When embedded in another layer, it appears the config.changed() helper doesn't telegraph that the configuration variable "has changed" on first run. This lent itself to a scenario where the CA was never getting generated after the TLS layer was embedded in a charm. Reworking to using synthetic states appears to have corrected this