Adds method to force regeneration of certs
Introduces the 'tls.regenerate_certificates' state, which allows consumer layers to reconfigure the primary ssl configuration certificate (on the master)

This has a side-effect of removing the flat file PKI database that is initially generated during TLS setup. This is unfortunate, but is a work-around to conform to expectations of the leader.

Follower behavior should be unchanged with this behavior.