Add the 'client certificate available' state
I need to key off an action when the client certificate is available.

This change is necessary as binding on 'server certificate available' will be triggered on peer units that do not have the client certificate.  As a leader i can distribute this key, but only when I know it is there.